Refactor signup methods to use same code
