+++ api/openapi_server/controllers/admin_controller.py
@@ -1,9 +1,12 @@
 
 import connexion
+from sqlalchemy.exc import IntegrityError
+from flask import session, current_app
 
 from openapi_server.controllers import auth_controller
-from flask import session, current_app
 from openapi_server.exceptions import AuthError
+from openapi_server.models.database import DataAccessLayer, User
+import botocore
 
 def initial_sign_in_reset_password(): 
     """Sets initial password.
@@ -58,3 +61,42 @@ def initial_sign_in_reset_password():
         'token': access_token,
         'user': user
     }
+
+def removeUser(body: dict):
+    '''
+    Remove a user from connected database and AWS Cognito user pool.
+    This method is only available to admin users.
+
+    body should contain the email of the user to be removed, e.g.:
+    {
+        "email": "EMAIL_TO_REMOVE"
+    }
+    '''   
+    with DataAccessLayer.session() as session:
+        user = session.query(User).filter_by(email=body['email']).first()
+        if user:
+            session.delete(user)
+        try:
+            session.commit()
+        except IntegrityError:
+            session.rollback()
+            # Since we're deleting, an IntegrityError might indicate a different problem
+            # Adjust the error message accordingly
+            raise AuthError({
+                "message": "Could not delete the user due to a database integrity constraint."
+            }, 422)
+        
+    try:
+        response = current_app.boto_client.admin_delete_user(
+            UserPoolId=current_app.config['COGNITO_USER_POOL_ID'],
+            Username=body['email']
+        )
+        return response
+    except botocore.exceptions.ClientError as error:
+        match error.response['Error']['Code']:
+            case 'UserNotFoundException':
+                msg = "User not found. Could not delete user."
+                raise AuthError({"message": msg}, 400)
+            case _:
+                msg = error.response['Error']['Message']
+                raise AuthError({"message": msg}, 500)
